Fix circular references
